requirements.txt
# requirements.txt
django>=3.2
djangorestframework>=3.12
djangorestframework-simplejwt>=5.2
pytest
pytest-django
twine